Move XCTestHelpers to TestCaseAccessibilityRule.swift
it's only used in that file
